+++ examples/Main.hs
@@ -8,7 +8,9 @@ module Main(
 import Hasklepias
 import ExampleEvents
 import ExampleFeatures1
+import ExampleFeatures2
 import Test.Hspec ( hspec )
 
 main :: IO ()
-main = hspec exampleFeatures1Spec
+main = hspec $ do exampleFeatures1Spec
+                  exampleFeatures2Spec